Tutorial mocks transferred to the cmd package
