+++ pkg/cmd/mocks_test.go
@@ -11,6 +11,7 @@ import (
 	"github.com/ZupIT/ritchie-cli/pkg/credential"
 	"github.com/ZupIT/ritchie-cli/pkg/formula"
 	"github.com/ZupIT/ritchie-cli/pkg/rcontext"
+	"github.com/ZupIT/ritchie-cli/pkg/rtutorial"
 )
 
 type inputTextMock struct{}
@@ -347,4 +348,28 @@ type InputMultilineMock struct{}
 
 func (InputMultilineMock) MultiLineText(name string, required bool) (string, error) {
 	return "username=ritchie", nil
+}
+
+type TutorialSetterMock struct{}
+
+func (TutorialSetterMock) Set(tutorial string) (rtutorial.TutorialHolder, error) {
+	return rtutorial.TutorialHolder{}, nil
+}
+
+type TutorialFinderMock struct{}
+
+func (TutorialFinderMock) Find() (rtutorial.TutorialHolder, error) {
+	return rtutorial.TutorialHolder{}, nil
+}
+
+type TutorialFindSetterMock struct{}
+
+func (TutorialFindSetterMock) Find() (rtutorial.TutorialHolder, error) {
+	f := TutorialFinderMock{}
+	return f.Find()
+}
+
+func (TutorialFindSetterMock) Set(tutorial string) (rtutorial.TutorialHolder, error) {
+	s := TutorialSetterMock{}
+	return s.Set(tutorial)
 }
