+++ pkg/cmd/tutorial_test.go
@@ -3,12 +3,10 @@ package cmd
 import (
 	"os"
 	"testing"
-
-	"github.com/ZupIT/ritchie-cli/pkg/mock"
 )
 
 func TestNewTutorialCmd(t *testing.T) {
-	cmd := NewTutorialCmd(os.TempDir(), inputListMock{}, mock.TutorialFindSetterMock{})
+	cmd := NewTutorialCmd(os.TempDir(), inputListMock{}, TutorialFindSetterMock{})
 	cmd.PersistentFlags().Bool("stdin", false, "input by stdin")
 
 	if cmd == nil {
